refactor(FE): `useDetachWallet` hook
